Structured implementation, incorporated recent extensions, started documentation
